+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 * The Doc Browser will now not change the menu when refreshing.
+* Fixes an issue where URLs in the doc browser would display JSON.
 
 ## 0.11.1